feat: :art: change achievement.d.ts to profile.d.ts
