+++ frontend/src/pages/ProfilePage/AchievementTooltip.tsx
@@ -1,4 +1,4 @@
-import { AchievementType } from 'types/achievement';
+import { AchievementType } from 'types/profile';
 
 interface Props {
   achieve: AchievementType;
