--- frontend/src/pages/ProfilePage/MatchHistory.tsx
@@ -1,7 +1,0 @@
-interface Props {
-  userId: number;
-}
-
-export const MatchHistory = ({ userId }: Props) => {
-  return <div>Match History userID: {userId}</div>;
-};
