--- frontend/src/types/achievement.d.ts
@@ -1,7 +1,0 @@
-export interface AchievementType {
-  id: number;
-  name: string;
-  description: string;
-  image: string;
-  achievedAt: string;
-}
