+++ frontend/src/types/profile.d.ts
@@ -1,0 +1,15 @@
+export interface AchievementType {
+  id: number;
+  name: string;
+  description: string;
+  image: string;
+  achievedAt: string;
+}
+
+export interface MatchHistoryType {
+  id: number;
+  name: string;
+  description: string;
+  image: string;
+  achievedAt: string;
+}
